+++ pineappl_cli/src/export.rs
@@ -157,6 +157,7 @@ impl Subcommand for Opts {
             let reference_results = helpers::convolve(
                 &grid,
                 &mut conv_funs,
+                &self.conv_funs.conv_types,
                 &orders,
                 &[],
                 &[],