+++ pineappl_cli/src/pull.rs
@@ -70,6 +70,7 @@ impl Subcommand for Opts {
                 Ok::<_, Error>(helpers::convolve(
                     &grid,
                     funs,
+                    &self.conv_funs1.conv_types,
                     &self.orders,
                     &[],
                     &[],
@@ -88,6 +89,7 @@ impl Subcommand for Opts {
                 Ok::<_, Error>(helpers::convolve(
                     &grid,
                     funs,
+                    &self.conv_funs2.conv_types,
                     &self.orders,
                     &[],
                     &[],
@@ -140,8 +142,8 @@ impl Subcommand for Opts {
             };
 
             let channel_results =
-                |member: Option<usize>, pdfset: &mut [Vec<Pdf>], set: &PdfSet| -> Vec<f64> {
-                    if let Some(member) = member {
+                |conv_funs: &ConvFuns, pdfset: &mut [Vec<Pdf>], set: &PdfSet| -> Vec<f64> {
+                    if let Some(member) = conv_funs.members[self.pull_from] {
                         (0..grid.channels().len())
                             .map(|channel| {
                                 let mut channel_mask = vec![false; grid.channels().len()];
@@ -149,6 +151,7 @@ impl Subcommand for Opts {
                                 match helpers::convolve(
                                     &grid,
                                     &mut pdfset[member],
+                                    &conv_funs.conv_types,
                                     &self.orders,
                                     &[bin],
                                     &channel_mask,
@@ -174,6 +177,7 @@ impl Subcommand for Opts {
                                         match helpers::convolve(
                                             &grid,
                                             fun,
+                                            &conv_funs.conv_types,
                                             &self.orders,
                                             &[bin],
                                             &channel_mask,
@@ -208,16 +212,8 @@ impl Subcommand for Opts {
             let mut pull_tuples = if self.limit == 0 {
                 Vec::new()
             } else {
-                let channel_results1 = channel_results(
-                    self.conv_funs1.members[self.pull_from],
-                    &mut conv_funs1,
-                    &set1,
-                );
-                let channel_results2 = channel_results(
-                    self.conv_funs2.members[self.pull_from],
-                    &mut conv_funs2,
-                    &set2,
-                );
+                let channel_results1 = channel_results(&self.conv_funs1, &mut conv_funs1, &set1);
+                let channel_results2 = channel_results(&self.conv_funs2, &mut conv_funs2, &set2);
 
                 let pull_tuples: Vec<_> = channel_results2
                     .iter()